Added confirmation pop up to delete post

+++ frontend/src/components/DeletePostButton.jsx
@@ -1,10 +1,19 @@
+import { useConfirm } from "material-ui-confirm";
 
 const DeletePostId = ({post_id, DeletePostId, UpdatePost, isYours}) =>{
-const token = localStorage.getItem("token");
+  const confirm = useConfirm()
+
+  const token = localStorage.getItem("token");
     return (
-      <button onClick={() => { 
-        DeletePostId(token, post_id, isYours)
-        setTimeout(() => {UpdatePost(Math.random())}, 200)
+      <button onClick={() => {
+        confirm({ description: "Post will be permanently deleted!" })
+          .then(() => {
+            DeletePostId(token, post_id, isYours)
+            setTimeout(() => {UpdatePost(Math.random())}, 200)
+          })
+          .catch(() => {
+            console.log("Post delete cancelled")
+          })
         }}>Delete Post</button>
   )
